+++ test/e2e/content.e2e.spec.js
@@ -49,7 +49,10 @@ describe('content (E2E):', function() {
             var mockExp = {
                 id: "e2e-pubget1",
                 title: "test experience",
+                data: [ { data: { foo: 'bar' }, versionId: 'a5e744d0' } ],
                 access: "public",
+                user: 'e2e-user',
+                org: 'e2e-org',
                 status: "active"
             };
             testUtils.resetCollection('experiences', mockExp).done(done);
@@ -65,6 +68,10 @@ describe('content (E2E):', function() {
                 expect(resp.body._id).not.toBeDefined();
                 expect(resp.body.id).toBe("e2e-pubget1");
                 expect(resp.body.title).toBe("test experience");
+                expect(resp.body.data).toEqual({foo: 'bar'});
+                expect(resp.body.user).not.toBeDefined();
+                expect(resp.body.org).not.toBeDefined();
+                expect(resp.body.versionId).toBe('a5e744d0');
                 done();
             }).catch(function(error) {
                 expect(error).not.toBeDefined();
@@ -127,21 +134,24 @@ describe('content (E2E):', function() {
                     title: "test experience",
                     access: "public",
                     status: "inactive",
-                    user: "e2e-user"
+                    user: "e2e-user",
+                    org: 'e2e-org'
                 },
                 {
                     id: "e2e-getid2",
                     title: "test experience",
                     access: "private",
                     status: "active",
-                    user: "not-e2e-user"
+                    user: "not-e2e-user",
+                    org: 'not-e2e-org'
                 },
                 {
                     id: "e2e-getid3",
                     title: "test experience",
                     access: "public",
                     status: "inactive",
-                    user: "not-e2e-user"
+                    user: "not-e2e-user",
+                    org: 'not-e2e-org'
                 }
             ];
             testUtils.resetCollection('experiences', mockExps).done(done);
@@ -158,6 +168,8 @@ describe('content (E2E):', function() {
                 expect(resp.body._id).not.toBeDefined();
                 expect(resp.body.id).toBe("e2e-getid1");
                 expect(resp.body.title).toBe("test experience");
+                expect(resp.body.user).toBe('e2e-user');
+                expect(resp.body.org).toBe('e2e-org');
                 done();
             }).catch(function(error) {
                 expect(error).not.toBeDefined();
@@ -263,8 +275,8 @@ describe('content (E2E):', function() {
                 expect(resp.body instanceof Array).toBeTruthy('body is array');
                 expect(resp.body.length).toBe(2);
                 expect(resp.body[0]._id).not.toBeDefined();
-                expect(resp.body[1]._id).not.toBeDefined();
                 expect(resp.body[0].id).toBe("e2e-getquery1");
+                expect(resp.body[1]._id).not.toBeDefined();
                 expect(resp.body[1].id).toBe("e2e-getquery3");
                 done();
             }).catch(function(error) {
@@ -420,6 +432,7 @@ describe('content (E2E):', function() {
         beforeEach(function(done) {
             mockExp = {
                 title: 'testExp',
+                data: { foo: 'bar' },
                 org: 'e2e-org'
             };
             testUtils.resetCollection('experiences').done(done);
@@ -437,6 +450,8 @@ describe('content (E2E):', function() {
                 expect(resp.body._id).not.toBeDefined();
                 expect(resp.body.id).toBeDefined();
                 expect(resp.body.title).toBe("testExp");
+                expect(resp.body.data).toEqual({foo: 'bar'});
+                expect(resp.body.versionId).toBe('a5e744d0');
                 expect(resp.body.user).toBe("e2e-user");
                 expect(resp.body.org).toBe("e2e-org");
                 expect(resp.body.created).toBeDefined();
@@ -498,6 +513,7 @@ describe('content (E2E):', function() {
             mockExps = [
                 {
                     id: "e2e-put1",
+                    data: [ { data: { foo: 'bar' }, versionId: 'a5e744d0' } ],
                     tag: "origTag",
                     status: "active",
                     access: "public",
@@ -524,12 +540,36 @@ describe('content (E2E):', function() {
             testUtils.qRequest('put', options).then(function(resp) {
                 expect(resp.response.statusCode).toBe(200);
                 updatedExp = resp.body;
-                expect(updatedExp).not.toEqual(mockExps);
+                expect(updatedExp).not.toEqual(mockExps[0]);
                 expect(updatedExp).toBeDefined();
                 expect(updatedExp._id).not.toBeDefined();
                 expect(updatedExp.id).toBe('e2e-put1');
                 expect(updatedExp.tag).toBe('newTag');
                 expect(updatedExp.user).toBe('e2e-user');
+                expect(updatedExp.versionId).toBe('a5e744d0');
+                expect(new Date(updatedExp.created)).toEqual(now);
+                expect(new Date(updatedExp.lastUpdated)).toBeGreaterThan(now);
+                done();
+            }).catch(function(error) {
+                expect(error).not.toBeDefined();
+                done();
+            });
+        });
+        
+        it('should properly update the data and versionId together', function(done) {
+            var options = {
+                url: config.contentUrl + '/content/experience/e2e-put1',
+                jar: cookieJar,
+                json: { data: { foo: 'baz' } }
+            };
+            testUtils.qRequest('put', options).then(function(resp) {
+                expect(resp.response.statusCode).toBe(200);
+                updatedExp = resp.body;
+                expect(updatedExp).not.toEqual(mockExps[0]);
+                expect(updatedExp).toBeDefined();
+                expect(updatedExp._id).not.toBeDefined();
+                expect(updatedExp.data).toEqual({foo: 'baz'});
+                expect(updatedExp.versionId).toBe('4c5c9754');
                 expect(new Date(updatedExp.created)).toEqual(now);
                 expect(new Date(updatedExp.lastUpdated)).toBeGreaterThan(now);
                 done();